LPS-58791 Add sample implementation to Bookmarks portlet
