+++ modules/apps/bookmarks/bookmarks-service/src/com/liferay/bookmarks/lar/BookmarksEntryStagedModelDataHandler.java
@@ -19,9 +19,7 @@ import com.liferay.bookmarks.model.BookmarksFolder;
 import com.liferay.bookmarks.model.BookmarksFolderConstants;
 import com.liferay.bookmarks.service.BookmarksEntryLocalServiceUtil;
 import com.liferay.exportimport.lar.BaseStagedModelDataHandler;
-import com.liferay.portal.kernel.dao.orm.QueryUtil;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.trash.TrashHandler;
+import com.liferay.exportimport.stagedmodel.repository.StagedModelRepository;
 import com.liferay.portal.kernel.util.MapUtil;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.service.ServiceContext;
@@ -29,12 +27,11 @@ import com.liferay.portlet.exportimport.lar.ExportImportPathUtil;
 import com.liferay.portlet.exportimport.lar.PortletDataContext;
 import com.liferay.portlet.exportimport.lar.StagedModelDataHandler;
 import com.liferay.portlet.exportimport.lar.StagedModelDataHandlerUtil;
-import com.liferay.portlet.exportimport.lar.StagedModelModifiedDateComparator;
 
-import java.util.List;
 import java.util.Map;
 
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Mate Thurzo
@@ -45,41 +42,6 @@ public class BookmarksEntryStagedModelDataHandler
 	extends BaseStagedModelDataHandler<BookmarksEntry> {
 
 	public static final String[] CLASS_NAMES = {BookmarksEntry.class.getName()};
-
-	@Override
-	public void deleteStagedModel(BookmarksEntry entry) throws PortalException {
-		BookmarksEntryLocalServiceUtil.deleteEntry(entry);
-	}
-
-	@Override
-	public void deleteStagedModel(
-			String uuid, long groupId, String className, String extraData)
-		throws PortalException {
-
-		BookmarksEntry entry = fetchStagedModelByUuidAndGroupId(uuid, groupId);
-
-		if (entry != null) {
-			deleteStagedModel(entry);
-		}
-	}
-
-	@Override
-	public BookmarksEntry fetchStagedModelByUuidAndGroupId(
-		String uuid, long groupId) {
-
-		return BookmarksEntryLocalServiceUtil.
-			fetchBookmarksEntryByUuidAndGroupId(uuid, groupId);
-	}
-
-	@Override
-	public List<BookmarksEntry> fetchStagedModelsByUuidAndCompanyId(
-		String uuid, long companyId) {
-
-		return BookmarksEntryLocalServiceUtil.
-			getBookmarksEntriesByUuidAndCompanyId(
-				uuid, companyId, QueryUtil.ALL_POS, QueryUtil.ALL_POS,
-				new StagedModelModifiedDateComparator<BookmarksEntry>());
-	}
 
 	@Override
 	public String[] getClassNames() {
@@ -179,24 +141,21 @@ public class BookmarksEntryStagedModelDataHandler
 	}
 
 	@Override
-	protected void doRestoreStagedModel(
-			PortletDataContext portletDataContext, BookmarksEntry entry)
-		throws Exception {
-
-		long userId = portletDataContext.getUserId(entry.getUserUuid());
-
-		BookmarksEntry existingEntry = fetchStagedModelByUuidAndGroupId(
-			entry.getUuid(), portletDataContext.getScopeGroupId());
-
-		if ((existingEntry == null) || !existingEntry.isInTrash()) {
-			return;
-		}
-
-		TrashHandler trashHandler = existingEntry.getTrashHandler();
-
-		if (trashHandler.isRestorable(existingEntry.getEntryId())) {
-			trashHandler.restoreTrashEntry(userId, existingEntry.getEntryId());
-		}
+	protected StagedModelRepository<BookmarksEntry> getStagedModelRepository() {
+		return _stagedModelRepository;
 	}
+
+	@Reference(
+		target =
+			"(model.class.name=com.liferay.bookmarks.model.BookmarksEntry)",
+		unbind = "-"
+	)
+	protected void setStagedModelRepository(
+		StagedModelRepository<BookmarksEntry> stagedModelRepository) {
+
+		_stagedModelRepository = stagedModelRepository;
+	}
+
+	private StagedModelRepository<BookmarksEntry> _stagedModelRepository;
 
 }
